Move completion commands to a different binary
