--- packages/hh/bin/index.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const { main } = require("../dist/src")
-
-main().catch(console.error)
